rename file name to useCustomerProductLists
