+++ packages/pwa/app/pages/cart/partials/cart-secondary-button-group.test.js
@@ -69,8 +69,8 @@ jest.mock('../../../commerce-api/hooks/useBasket', () => {
     }
 })
 
-jest.mock('../../../commerce-api/hooks/useCustomerProductList', () => {
-    const originalModule = jest.requireActual('../../../commerce-api/hooks/useCustomerProductList')
+jest.mock('../../../commerce-api/hooks/useCustomerProductLists', () => {
+    const originalModule = jest.requireActual('../../../commerce-api/hooks/useCustomerProductLists')
     const useCustomerProductLists = originalModule.default
     return () => {
         const customerProductLists = useCustomerProductLists()
